+++ druid/src/core.rs
@@ -20,8 +20,8 @@ use tracing::{trace, trace_span, warn};
 use crate::bloom::Bloom;
 use crate::command::sys::{CLOSE_WINDOW, SUB_WINDOW_HOST_TO_PARENT, SUB_WINDOW_PARENT_TO_HOST};
 use crate::commands::SCROLL_TO_VIEW;
-use crate::contexts::ContextState;
-use crate::kurbo::{Affine, Insets, Point, Rect, Shape, Size, Vec2};
+use crate::contexts::{ChangeCtx, ContextState};
+use crate::kurbo::{Affine, Insets, Point, Rect, Shape, Size};
 use crate::sub_window::SubWindowUpdate;
 use crate::{
     ArcStr, BoxConstraints, Color, Command, Cursor, Data, Env, Event, EventCtx, InternalEvent,
@@ -80,7 +80,10 @@ pub struct WidgetState {
     /// The origin of the child in the parent's coordinate space; together with
     /// `size` these constitute the child's layout rect.
     origin: Point,
-    /// The origin of the parent in the window coordinate space;
+    /// The origin of the parent in the window coordinate space.
+    ///
+    /// Note that this value can be stale after `set_origin` was called
+    /// but before `LifeCycle::ViewContextChanged` has fully propagated.
     pub(crate) parent_window_origin: Point,
     /// A flag used to track and debug missing calls to set_origin.
     is_expecting_set_origin_call: bool,
@@ -98,11 +101,6 @@ pub struct WidgetState {
 
     // The region that needs to be repainted, relative to the widget's bounds.
     pub(crate) invalid: Region,
-
-    // The part of this widget that is visible on the screen is offset by this
-    // much. This will be non-zero for widgets that are children of `Scroll`, or
-    // similar, and it is used for propagating invalid regions.
-    pub(crate) viewport_offset: Vec2,
 
     // TODO: consider using bitflags for the booleans.
     // `true` if a descendent of this widget changed its disabled state and should receive
@@ -126,8 +124,16 @@ pub struct WidgetState {
 
     pub(crate) needs_layout: bool,
 
-    /// Because of some scrolling or something, `parent_window_origin` needs to be updated.
-    pub(crate) needs_window_origin: bool,
+    /// Some of our children have the `view_context_changed` flag set.
+    pub(crate) children_view_context_changed: bool,
+
+    /// indicate that the [`ViewContext`] changed.
+    ///
+    /// When this flag is set, the associated widget will receive a `LifeCycle::ViewContextChanged`
+    /// event.
+    ///
+    /// [`ViewContext`]: crate::ViewContext
+    pub(crate) view_context_changed: bool,
 
     /// Any descendant is active.
     has_active: bool,
@@ -253,11 +259,11 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
     /// This is soft-deprecated; you should use [`set_origin`] instead for new code.
     ///
     /// [`set_origin`]: WidgetPod::set_origin
-    pub fn set_layout_rect(&mut self, ctx: &mut LayoutCtx, data: &T, env: &Env, layout_rect: Rect) {
+    pub fn set_layout_rect(&mut self, ctx: &mut LayoutCtx, layout_rect: Rect) {
         if layout_rect.size() != self.state.size {
             warn!("set_layout_rect passed different size than returned by layout method");
         }
-        self.set_origin(ctx, data, env, layout_rect.origin());
+        self.set_origin(ctx, layout_rect.origin());
     }
 
     /// Set the origin of this widget, in the parent's coordinate space.
@@ -266,29 +272,30 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
     /// its own [`Widget::layout`] implementation, and then call `set_origin` to
     /// position those children.
     ///
+    /// The changed origin won't be fully in effect until [`LifeCycle::ViewContextChanged`] has
+    /// finished propagating. Specifically methods that depend on the widget's origin in relation
+    /// to the window will return stale results during the period after calling `set_origin` but
+    /// before [`LifeCycle::ViewContextChanged`] has finished propagating.
+    ///
+    /// The widget container can also call `set_origin` from other context, but calling `set_origin`
+    /// after the widget received [`LifeCycle::ViewContextChanged`] and before the next event results
+    /// in an inconsistent state of the widget tree.
+    ///
     /// The child will receive the [`LifeCycle::Size`] event informing them of the final [`Size`].
     ///
     /// [`Widget::layout`]: trait.Widget.html#tymethod.layout
     /// [`Rect`]: struct.Rect.html
     /// [`Size`]: struct.Size.html
     /// [`LifeCycle::Size`]: enum.LifeCycle.html#variant.Size
-    pub fn set_origin(&mut self, ctx: &mut LayoutCtx, data: &T, env: &Env, origin: Point) {
-        self.state.origin = origin;
+    /// [`LifeCycle::ViewContextChanged`]: enum.LifeCycle.html#variant.ViewContextChanged
+    pub fn set_origin(&mut self, ctx: &mut impl ChangeCtx, origin: Point) {
         self.state.is_expecting_set_origin_call = false;
-        let layout_rect = self.layout_rect();
 
-        // if the widget has moved, it may have moved under the mouse, in which
-        // case we need to handle that.
-        if WidgetPod::set_hot_state(
-            &mut self.inner,
-            &mut self.state,
-            ctx.state,
-            layout_rect,
-            ctx.mouse_pos,
-            data,
-            env,
-        ) {
-            ctx.widget_state.merge_up(&mut self.state);
+        if origin != self.state.origin {
+            self.state.origin = origin;
+            self.state.view_context_changed = true;
+            // Instead of calling merge_up we directly propagate the flag we care about, to gain performance
+            ctx.state().widget_state.children_view_context_changed = true;
         }
     }
 
@@ -303,31 +310,6 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
     /// [`set_origin`]: WidgetPod::set_origin
     pub fn layout_rect(&self) -> Rect {
         self.state.layout_rect()
-    }
-
-    /// Set the viewport offset.
-    ///
-    /// This is relevant only for children of a scroll view (or similar). It must
-    /// be set by the parent widget whenever it modifies the position of its child
-    /// while painting it and propagating events. As a rule of thumb, you need this
-    /// if and only if you `Affine::translate` the paint context before painting
-    /// your child. For an example, see the implementation of [`Scroll`].
-    ///
-    /// [`Scroll`]: widget/struct.Scroll.html
-    pub fn set_viewport_offset(&mut self, offset: Vec2) {
-        if offset != self.state.viewport_offset {
-            self.state.needs_window_origin = true;
-        }
-        self.state.viewport_offset = offset;
-    }
-
-    /// The viewport offset.
-    ///
-    /// This will be the same value as set by [`set_viewport_offset`].
-    ///
-    /// [`set_viewport_offset`]: #method.viewport_offset
-    pub fn viewport_offset(&self) -> Vec2 {
-        self.state.viewport_offset
     }
 
     /// Get the widget's paint [`Rect`].
@@ -382,41 +364,41 @@ impl<T, W: Widget<T>> WidgetPod<T, W> {
         self.state.baseline_offset
     }
 
-    /// Determines if the provided `mouse_pos` is inside `rect`
+    /// Determines if the provided `mouse_pos` is inside the widget's `layout_rect`
     /// and if so updates the hot state and sends `LifeCycle::HotChanged`.
     ///
     /// Returns `true` if the hot state changed.
     ///
-    /// The provided `child_state` should be merged up if this returns `true`.
+    /// The WidgetPod should merge up its state when `true` is returned.
     fn set_hot_state(
-        child: &mut W,
-        child_state: &mut WidgetState,
+        &mut self,
         state: &mut ContextState,
-        rect: Rect,
         mouse_pos: Option<Point>,
         data: &T,
         env: &Env,
     ) -> bool {
-        let had_hot = child_state.is_hot;
-        child_state.is_hot = match mouse_pos {
+        let rect = self.layout_rect();
+        let had_hot = self.state.is_hot;
+        self.state.is_hot = match mouse_pos {
             Some(pos) => rect.winding(pos) != 0,
             None => false,
         };
-        if had_hot != child_state.is_hot {
+        if had_hot != self.state.is_hot {
             trace!(
                 "Widget {:?}: set hot state to {}",
-                child_state.id,
-                child_state.is_hot
+                self.state.id,
+                self.state.is_hot
             );
 
-            let hot_changed_event = LifeCycle::HotChanged(child_state.is_hot);
+            let hot_changed_event = LifeCycle::HotChanged(self.state.is_hot);
             let mut child_ctx = LifeCycleCtx {
                 state,
-                widget_state: child_state,
+                widget_state: &mut self.state,
             };
             // We add a span so that inner logs are marked as being in a lifecycle pass
+            let widget = &mut self.inner;
             trace_span!("lifecycle")
-                .in_scope(|| child.lifecycle(&mut child_ctx, &hot_changed_event, data, env));
+                .in_scope(|| widget.lifecycle(&mut child_ctx, &hot_changed_event, data, env));
             // if hot changes and we're showing widget ids, always repaint
             if env.get(Env::DEBUG_WIDGET_ID) {
                 child_ctx.request_paint();
@@ -450,22 +432,22 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
             state: ctx.state,
             z_ops: Vec::new(),
             region: ctx.region.clone(),
-            widget_state: &self.state,
+            widget_state: &mut self.state,
             depth: ctx.depth,
         };
         self.inner.paint(&mut inner_ctx, data, env);
 
+        ctx.z_ops.append(&mut inner_ctx.z_ops);
+
         let debug_ids = inner_ctx.is_hot() && env.get(Env::DEBUG_WIDGET_ID);
         if debug_ids {
             // this also draws layout bounds
-            self.debug_paint_widget_ids(&mut inner_ctx, env);
+            self.debug_paint_widget_ids(ctx, env);
         }
 
         if !debug_ids && env.get(Env::DEBUG_PAINT) {
-            self.debug_paint_layout_bounds(&mut inner_ctx, env);
+            self.debug_paint_layout_bounds(ctx, env);
         }
-
-        ctx.z_ops.append(&mut inner_ctx.z_ops);
     }
 
     /// Paint the widget, translating it by the origin of its layout rectangle.
@@ -571,22 +553,13 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
         }
 
         self.state.needs_layout = false;
-        self.state.needs_window_origin = false;
         self.state.is_expecting_set_origin_call = true;
 
-        //TODO: this does not work!
-        // self.layout_rect().origin().to_vec2() + self.viewport_offset() is the old position which
-        // changes after set origin. Therefore changing hot state must happen after the root widget
-        // set its origin.
-        let child_mouse_pos = ctx
-            .mouse_pos
-            .map(|pos| pos - self.layout_rect().origin().to_vec2() + self.viewport_offset());
         let prev_size = self.state.size;
 
         let mut child_ctx = LayoutCtx {
             widget_state: &mut self.state,
             state: ctx.state,
-            mouse_pos: child_mouse_pos,
         };
 
         let new_size = self.inner.layout(&mut child_ctx, bc, data, env);
@@ -685,15 +658,7 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
         let recurse = match event {
             Event::Internal(internal) => match internal {
                 InternalEvent::MouseLeave => {
-                    let hot_changed = WidgetPod::set_hot_state(
-                        &mut self.inner,
-                        &mut self.state,
-                        ctx.state,
-                        rect,
-                        None,
-                        data,
-                        env,
-                    );
+                    let hot_changed = self.set_hot_state(ctx.state, None, data, env);
                     had_active || hot_changed
                 }
                 InternalEvent::TargetedCommand(cmd) => {
@@ -743,11 +708,8 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 ctx.is_root
             }
             Event::MouseDown(mouse_event) => {
-                WidgetPod::set_hot_state(
-                    &mut self.inner,
-                    &mut self.state,
+                self.set_hot_state(
                     ctx.state,
-                    rect,
                     if !ctx.is_handled {
                         Some(mouse_event.pos)
                     } else {
@@ -766,11 +728,8 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 }
             }
             Event::MouseUp(mouse_event) => {
-                WidgetPod::set_hot_state(
-                    &mut self.inner,
-                    &mut self.state,
+                self.set_hot_state(
                     ctx.state,
-                    rect,
                     if !ctx.is_handled {
                         Some(mouse_event.pos)
                     } else {
@@ -789,11 +748,8 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 }
             }
             Event::MouseMove(mouse_event) => {
-                let hot_changed = WidgetPod::set_hot_state(
-                    &mut self.inner,
-                    &mut self.state,
+                let hot_changed = self.set_hot_state(
                     ctx.state,
-                    rect,
                     if !ctx.is_handled {
                         Some(mouse_event.pos)
                     } else {
@@ -815,11 +771,8 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 }
             }
             Event::Wheel(mouse_event) => {
-                WidgetPod::set_hot_state(
-                    &mut self.inner,
-                    &mut self.state,
+                self.set_hot_state(
                     ctx.state,
-                    rect,
                     if !ctx.is_handled {
                         Some(mouse_event.pos)
                     } else {
@@ -1012,10 +965,27 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                         _ => false,
                     }
                 }
-                InternalLifeCycle::ParentWindowOrigin => {
-                    self.state.parent_window_origin = ctx.widget_state.window_origin();
-                    self.state.needs_window_origin = false;
-                    true
+                InternalLifeCycle::RouteViewContextChanged(view_context) => {
+                    if self.state.view_context_changed {
+                        self.lifecycle(
+                            ctx,
+                            &LifeCycle::ViewContextChanged(*view_context), // we dont aplly our origin. This happens in self.lifecycle(...).
+                            data,
+                            env,
+                        );
+
+                        return;
+                    } else if self.state.children_view_context_changed {
+                        extra_event = Some(LifeCycle::Internal(
+                            InternalLifeCycle::RouteViewContextChanged(
+                                view_context.for_child_widget(self.state.origin),
+                            ),
+                        ));
+
+                        self.state.children_view_context_changed = false;
+                    }
+
+                    false
                 }
                 InternalLifeCycle::DebugRequestState { widget, state_cell } => {
                     if *widget == self.id() {
@@ -1099,6 +1069,19 @@ impl<T: Data, W: Widget<T>> WidgetPod<T, W> {
                 } else {
                     false
                 }
+            }
+            LifeCycle::ViewContextChanged(view_context) => {
+                extra_event = Some(LifeCycle::ViewContextChanged(
+                    view_context.for_child_widget(self.state.origin),
+                ));
+
+                self.set_hot_state(ctx.state, view_context.last_mouse_position, data, env);
+                self.state.parent_window_origin = view_context.window_origin;
+
+                self.state.children_view_context_changed = false;
+                self.state.view_context_changed = false;
+
+                false
             }
         };
 
@@ -1259,14 +1242,13 @@ impl WidgetState {
             is_expecting_set_origin_call: true,
             paint_insets: Insets::ZERO,
             invalid: Region::EMPTY,
-            viewport_offset: Vec2::ZERO,
             children_disabled_changed: false,
             ancestor_disabled: false,
             is_explicitly_disabled: false,
             baseline_offset: 0.0,
             is_hot: false,
             needs_layout: false,
-            needs_window_origin: false,
+            children_view_context_changed: false,
             is_active: false,
             has_active: false,
             has_focus: false,
@@ -1281,6 +1263,7 @@ impl WidgetState {
             sub_window_hosts: Vec::new(),
             is_explicitly_disabled_new: false,
             update_focus_chain: false,
+            view_context_changed: false,
         }
     }
 
@@ -1303,7 +1286,7 @@ impl WidgetState {
             .layout_rect()
             .with_origin(Point::ORIGIN)
             .inset(self.paint_insets);
-        let offset = child_state.layout_rect().origin().to_vec2() - child_state.viewport_offset;
+        let offset = child_state.layout_rect().origin().to_vec2();
         for &r in child_state.invalid.rects() {
             let r = (r + offset).intersect(clip);
             if r.area() != 0.0 {
@@ -1317,7 +1300,8 @@ impl WidgetState {
         child_state.invalid.clear();
 
         self.needs_layout |= child_state.needs_layout;
-        self.needs_window_origin |= child_state.needs_window_origin;
+        self.children_view_context_changed |=
+            child_state.children_view_context_changed | child_state.view_context_changed;
         self.request_anim |= child_state.request_anim;
         self.children_disabled_changed |= child_state.children_disabled_changed;
         self.children_disabled_changed |=
@@ -1376,7 +1360,7 @@ impl WidgetState {
     }
 
     pub(crate) fn window_origin(&self) -> Point {
-        self.parent_window_origin + self.origin.to_vec2() - self.viewport_offset
+        self.parent_window_origin + self.origin.to_vec2()
     }
 }